refactor message into own file
